Day 1: Refactor to use same function name with pattern matching
My code was not very idiomatic as I was defining different function names that were doing similar things.
In Elixir is more idiomatic to use same function name but with different pattern matching.
In this case I have defined a tuple with an atom that defines the type of the data passed in.